+++ src/web/src/components/Search/Detail/SearchDetail.jsx
@@ -295,7 +295,7 @@ const SearchDetail = ({
 
   const create = async ({ navigate, search: searchForCreate }) => {
     reset();
-    onCreate({ navigate, searchForCreate });
+    onCreate({ navigate, search: searchForCreate });
   };
 
   const openDiscoveryGraph = async (request) => {